Добавил обработку ошибок

+++ src/main/java/com/example/FinTrack/service/UserService.java
@@ -1,5 +1,6 @@
 package com.example.FinTrack.service;
 
+import com.example.FinTrack.exception.UserNotFoundException;
 import com.example.FinTrack.model.entity.User;
 import com.example.FinTrack.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -13,17 +14,25 @@ public class UserService {
         userRepository.save(user);
     }
     public boolean existsByUsername(String userName) {
-        return userRepository.existsByUserName(userName);
+        if(!userRepository.existsByUserName(userName)) {
+            throw new UserNotFoundException();
+        }
+        return true;
     }
     public User getByUserName(String userName) {
-        return userRepository.findByUserName(userName);
+        return userRepository.findByUserName(userName)
+                .orElseThrow(UserNotFoundException::new);
     }
 
     public User getByUserId(Long userId) {
-        return userRepository.findById(userId).orElseThrow();
+        return userRepository.findById(userId).orElseThrow(UserNotFoundException::new);
     }
 
     public void deleteUser(Long userId) {
-        userRepository.deleteById(userId);
+        if(!userRepository.existsById(userId)) {
+            userRepository.deleteById(userId);
+        } else {
+            throw new UserNotFoundException();
+        }
     }
 }
